カート機能 修正

+++ app/Http/Controllers/CartController.php
@@ -60,11 +60,14 @@ class CartController extends Controller
 
   public function add(Request $request)
   {
-    Cart::create([
-      'product_id' => $request->product_id,
-      'user_id' => Auth::user()->id,
-      'amount' => $request->amount
-    ]);
+    if($request->amount > 0) {
+      Cart::create([
+        'product_id' => $request->product_id,
+        'user_id' => Auth::user()->id,
+        'amount' => $request->amount
+      ]);
+    }
+
     return redirect('/cart');
   }
 
@@ -72,8 +75,10 @@ class CartController extends Controller
   {
     $user = Cart::select('user_id')->where('id', $request->cart_id)->first();
 
-    if(intval($user->user_id) === Auth::user()->id) {
-      Cart::destroy($request->cart_id);
+    if(is_object($user)) {
+      if(intval($user->user_id) === Auth::user()->id) {
+        Cart::destroy($request->cart_id);
+      }
     }
 
     return redirect('/cart');
